+++ packages/test/local-server-stress-tests/src/test/localServerStress.spec.ts
@@ -98,6 +98,6 @@ describe("Local Server Stress", () => {
 		// only: [99],
 		saveFailures,
 		// saveSuccesses,
-		skip: [93],
+		// skip: [93],
 	});
 });
